+++ mobile/android/android-components/docs/changelog.md
@@ -5,7 +5,6 @@ permalink: /changelog/
 ---
 
 # 5.0.0-SNAPSHOT  (In Development)
-
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v4.0.0...master)
 * [Milestone](https://github.com/mozilla-mobile/android-components/milestone/64?closed=1)
 * [Dependencies](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Dependencies.kt)
@@ -14,6 +13,9 @@ permalink: /changelog/
 
 * **All components**
   * Increased `compileSdkVersion` to 29 (Android Q)
+
+* **support-ktx**
+  * ⚠️ **This is a breaking behavior change**: `JSONArray.mapNotNull` is now an inline function, changing the behavior of the `return` keyword within its lambda.
 
 # 4.0.0