+++ mobile/android/android-components/docs/changelog.md
@@ -22,6 +22,18 @@ permalink: /changelog/
 
 * **support-ktx**
   * ⚠️ **This is a breaking behavior change**: `JSONArray.mapNotNull` is now an inline function, changing the behavior of the `return` keyword within its lambda.
+  
+# 4.0.1
+
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v4.0.0...v4.0.1)
+* [Dependencies](https://github.com/mozilla-mobile/android-components/blob/v4.0.1/buildSrc/src/main/java/Dependencies.kt)
+* [Gecko](https://github.com/mozilla-mobile/android-components/blob/v4.0.1/buildSrc/src/main/java/Gecko.kt)
+* [Configuration](https://github.com/mozilla-mobile/android-components/blob/v4.0.1/buildSrc/src/main/java/Config.kt)
+
+* **service-glean**
+  * Hyphens `-` are now allowed in labels for metrics.  See [1566764](https://bugzilla.mozilla.org/show_bug.cgi?id=1566764).
+  
+* Imported latest state of translations.
 
 # 4.0.0